refactor(RegistryV2): refacto all the things
