+++ front/src/form/registry/incomingTexs/shape.ts
@@ -17,7 +17,7 @@ import {
 import { Address } from "../common/Address";
 import { CompanySelector } from "../common/CompanySelector";
 import { EcoOrganismes } from "../common/EcoOrganismes";
-import { ParcelsVisualizer } from "../common/ParcelsVisualizer";
+import { ParcelsVisualizer } from "../common/ParcelsVisualizer/ParcelsVisualizer";
 import { ReportFor } from "../common/ReportFor";
 import { TransporterSelector } from "../common/TransporterSelector/TransporterSelector";
 import { WasteCodeSelector } from "../common/WasteCodeSelector";
